Fixing my mess with CardanoAPI
